% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scvelor.R
\name{scvelor}
\alias{scvelor}
\title{Wrapper for scVelo}
\usage{
scvelor(
  spliced,
  unspliced,
  output_anndata = FALSE,
  params = list(velocity = list(mode = "stochastic"))
)
}
\arguments{
\item{spliced, unspliced}{Count matrices with spliced and unspliced
counts. These matrices must have the same number of rows (representing
genes) and columns (representing cells).}

\item{output_anndata}{Logical, whether to return the AnnData object
with all estimated values.}

\item{params}{List of lists, allowing the user to provide arguments to
the functions in scVelo. The elements of \code{params} are named
after the respective scVelo function (e.g., \code{filter_and_normalize},
see full list above). Each element is a list, defining the arguments
to provide to the function. The only argument that must be defined
is \code{params$velocity$mode}, which has to be one of 'steady_state',
'stochastic' or 'dynamical'.}
}
\description{
This function provides a wrapper for running the scVelo pipeline
to estimate RNA velocities based on spliced and unspliced counts.
The following sequence of functions will be run, after the creation
of an AnnData object:
* pp.filter_and_normalize
* pp.moments
* tl.recover_dynamics (only for the dynamical model)
* tl.velocity
* tl.velocity_graph
* tl.velocity_pseudotime
* tl.latent_time (only for the dynamical model)
* tl.velocity_confidence
}
\references{
Bergen et al (2019)
}
\author{
Charlotte Soneson
}
